requirements.txt
fastapi
uvicorn
akshare
google-generativeai
python-dotenv
